Add script and route to login to Heroku (#100)
One of the big reasons we use web_git is to avoid teaching CLI stuff. Heroku CLI is much simpler to use than git, but since they are very much connected there should also be some web interface for the Heroku commands that are common.

This commit adds nav tabs to separate Git commands and Heroku commands. It also adds the dependency on expect—which seems to be the only reliable way to auto-login to Heroku CLI. In regards to this, the installer has been updated to account for the new dependency and also handle repeat runs.