Add config settings to control Heroku features (#142)
The output of the heroku whoami command occasionally caused some confusion, as the failure message was always present in the server log. Additionally, since we also don't plan to use Heroku for the foreseeable future, the Heroku features don't need to be enabled by default.

This  commit adds the ability to turn the Heroku features on/off in an initializer. The Heroku features are not enabled by default. This also hides and disables the Heroku nav tab and no longer call heroku whoami on every page load.